+++ python/ray/train/xgboost/__init__.py
@@ -1,9 +1,9 @@
+from ray.train.v1.xgboost._xgboost_utils import RayTrainReportCallback
+from ray.train.v1.xgboost.config import XGBoostConfig
+from ray.train.v1.xgboost.xgboost_checkpoint import XGBoostCheckpoint
+from ray.train.v1.xgboost.xgboost_predictor import XGBoostPredictor
+from ray.train.v1.xgboost.xgboost_trainer import XGBoostTrainer
 from ray.train.v2._internal.constants import is_v2_enabled
-from ray.train.xgboost._xgboost_utils import RayTrainReportCallback
-from ray.train.xgboost.config import XGBoostConfig
-from ray.train.xgboost.xgboost_checkpoint import XGBoostCheckpoint
-from ray.train.xgboost.xgboost_predictor import XGBoostPredictor
-from ray.train.xgboost.xgboost_trainer import XGBoostTrainer
 
 if is_v2_enabled():
     from ray.train.v2.xgboost.xgboost_trainer import XGBoostTrainer  # noqa: F811
